requirements-test.txt
apache-airflow-providers-mongo==4.0.0
apache-airflow-providers-amazon
boto3
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
apache-airflow==3.0.6
pandas>=1.5.0
pymongo>=4.0.0